Inital Commit + Base Camera Controls
Initial Commit of the project, contains extremely basic camera controls that changes where the player is looking if the mouse crosses specific thresholds of the screen, snapping to each cardinal direction. There is also a failsafe so that it only applies once every time the mouse is on the threshold.